webapp2_std_cloudstorage/requirements.txt
webapp2>=3.0.0b1
GoogleAppEngineCloudStorageClient
